+++ src/db_migrate.cpp
@@ -14,7 +14,7 @@
 namespace
 {
 
-constexpr auto version = std::string_view{"0.10.6"};
+constexpr auto version = std::string_view{"0.10.7"};
 
 auto print_help() -> void
 {